global: re-organize package structure
closes #80
