+++ pkg/validator/validator_test.go
@@ -6,7 +6,7 @@ REANA is free software; you can redistribute it and/or modify it
 under the terms of the MIT License; see LICENSE file for more details.
 */
 
-package validation
+package validator
 
 import (
 	"testing"
